+++ src/Modifiers.sol
@@ -4,7 +4,7 @@ pragma solidity ^0.8.2;
 import "./CampaignBase.sol";
 
 contract Modifiers is CampaignBase {
-    modifier onlyCreator(uint _campaignId) {
+    modifier onlyCreator(uint256 _campaignId) {
         require(msg.sender == campaigns[_campaignId].creator, "Unauthorised access");
         _;
     }
@@ -14,22 +14,22 @@ contract Modifiers is CampaignBase {
         _;
     }
 
-    modifier goalMet(uint _campaignId) {
+    modifier goalMet(uint256 _campaignId) {
         require(campaigns[_campaignId].pledged >= campaigns[_campaignId].goal, "Goal hasn't been met");
         _;
     }
 
-    modifier goalNotMet(uint _campaignId) {
+    modifier goalNotMet(uint256 _campaignId) {
         require(campaigns[_campaignId].pledged < campaigns[_campaignId].goal, "Goal has been met");
         _;
     }
 
-    modifier notClaimed(uint _campaignId) {
+    modifier notClaimed(uint256 _campaignId) {
         require(campaigns[_campaignId].claimed == false, "Funds have already been claimed");
         _;
     }
 
-    modifier fundsAvailable(uint _campaignId) {
+    modifier fundsAvailable(uint256 _campaignId) {
         require(campaigns[_campaignId].contributions[msg.sender] > 0, " ");
         _;
     }
